+++ udata/tests/helpers.py
@@ -1,4 +1,5 @@
 import os
+import traceback
 from collections.abc import Callable, Iterable
 from contextlib import contextmanager
 from datetime import datetime, timedelta
@@ -201,8 +202,10 @@ def assert500(response):
 
 def assert_command_ok(result):
     __tracebackhide__ = True
-    msg = "Command failed with exit code {0.exit_code} and output:\n{0.output}"
-    assert result.exit_code == 0, msg.format(result)
+    msg = f"Command failed with exit code {result.exit_code} and output:\n{result.output}"
+    if result.exception:
+        msg += "\n" + "".join(traceback.format_exception(result.exception))
+    assert result.exit_code == 0, msg
 
 
 def assert_urls_equal(url1, url2):